# service/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Studio 实验室管理系统 API",
        "title": "ScienceOL Studio API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "harveyque@outlook.com"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0",
            "url": "http://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "host": "localhost:48197",
    "basePath": "/api",
    "paths": {
...
        "/v1/lab": {
            "post": {
                "description": "创建一个新的实验室环境",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "创建实验室环境",
                "parameters": [
                    {
                        "description": "实验室环境创建请求",
                        "name": "lab",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/environment.LaboratoryEnvReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除一个实验室环境",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "删除实验室环境",
                "parameters": [
                    {
                        "description": "删除实验室请求",
                        "name": "lab",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/environment.DelLabReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "description": "更新一个已存在的实验室环境",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "更新实验室环境",
                "parameters": [
                    {
                        "description": "实验室环境更新请求",
                        "name": "lab",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/environment.UpdateEnvReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/invite/{lab_uuid}": {
            "post": {
                "description": "为实验室创建邀请链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "创建邀请链接",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实验室UUID",
                        "name": "lab_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/invite/{uuid}": {
            "get": {
                "description": "通过邀请链接加入实验室",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "接受邀请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邀请链接UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/list": {
            "get": {
                "description": "获取当前用户的所有实验室",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "获取实验室列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/lab/member/{lab_uuid}": {
            "get": {
                "description": "获取指定实验室的成员列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "获取实验室成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实验室UUID",
                        "name": "lab_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/member/{lab_uuid}/{member_uuid}": {
            "delete": {
                "description": "从实验室中删除一个成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "删除实验室成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实验室UUID",
                        "name": "lab_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "成员UUID",
                        "name": "member_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/lab/resource": {
            "post": {
                "description": "从边缘端创建实验室资源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "创建实验室资源",
                "parameters": [
                    {
                        "description": "资源请求",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/environment.ResourceReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/lab/{lab_uuid}": {
            "get": {
                "description": "获取单个实验室的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Laboratory"
                ],
                "summary": "获取实验室信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实验室UUID",
                        "name": "lab_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否包含成员列表",
                        "name": "with_member",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/definitions/code.ErrCode"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "common.Error": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "common.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/code.ErrCode"
                },
                "data": {},
                "error": {
                    "$ref": "#/definitions/common.Error"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "environment.DelLabReq": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "environment.LaboratoryEnvReq": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "environment.ResourceReq": {
            "type": "object"
        },
        "environment.UpdateEnvReq": {
            "type": "object",
            "required": [
                "uuid"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}